updated fusers

+++ src/main/java/de/uni_mannheim/informatik/wdi/datafusion/usecase/wdiproject/fusers/PopulationFuser.java
@@ -6,7 +6,7 @@ import de.uni_mannheim.informatik.wdi.datafusion.RecordGroup;
 import de.uni_mannheim.informatik.wdi.datafusion.conflictresolution.numeric.Median;
 import de.uni_mannheim.informatik.wdi.datafusion.usecase.wdiproject.FusableCity;
 
-public class PopulationFuser extends AttributeValueFuser<Integer, FusableCity> {
+public class PopulationFuser extends AttributeValueFuser<Double, FusableCity> {
 
 	public PopulationFuser() {
 		super(new Median<FusableCity>());
@@ -18,15 +18,15 @@ public class PopulationFuser extends AttributeValueFuser<Integer, FusableCity> {
 	}
 	
 	@Override
-	protected Integer getValue(FusableCity record) {
-		return record.getPopulation();
+	protected Double getValue(FusableCity record) {
+		return (double) record.getPopulation();
 	}
 
 	@Override
 	public void fuse(RecordGroup<FusableCity> group,
 			FusableCity fusedRecord) {
-		FusedValue<Integer, FusableCity> fused = getFusedValue(group);
-		fusedRecord.setPopulation(fused.getValue());
+		FusedValue<Double, FusableCity> fused = getFusedValue(group);
+		fusedRecord.setPopulation(fused.getValue().intValue());
 		fusedRecord.setAttributeProvenance(FusableCity.POPULATION, fused.getOriginalIds());
 	}
